# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.CustomerDTO:
    properties:
      delivery_type:
        type: string
      email:
        type: string
      email_notification:
        type: boolean
      father_name:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      payment_type:
        type: string
      phone:
        type: string
      telegram:
        type: string
      telegram_notification:
        type: boolean
      website_alias:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.UpdateCustomerProfileRequest:
    properties:
      delivery_type:
        type: string
      email_notification:
        type: boolean
      father_name:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      payment_type:
        type: string
      phone:
        type: string
      telegram:
        type: string
      telegram_notification:
        type: boolean
    type: object
...
host: localhost:8082
info:
  contact: {}
  title: Maestro
  version: "1.0"
paths:
...
  /customer/get-all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CustomerDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all customers in website TO ADMIN
      tags:
      - customer
...
  /customer/update-profile:
    patch:
      consumes:
      - application/json
      parameters:
      - description: customer new profile info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCustomerProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CustomerDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update customer profile
      tags:
      - customer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Token
    type: apiKey
swagger: "2.0"